we always want a sealed even if a del request for a draft is present and is newer than the sealed... This should never happen but we definantly want the sealed.

+++ source/org/martus/server/formirroring/MirroringRetriever.java
@@ -231,7 +231,7 @@ public class MirroringRetriever implements LoggerInterface
 		{
 			UniversalId uid = mirroringInfo.getUid();
 			DatabaseKey delKey = DeleteRequestRecord.getDelKey(uid);
-			if(store.doesBulletinDelRecordExist(delKey))
+			if(mirroringInfo.isDraft() && store.doesBulletinDelRecordExist(delKey))
 			{
 				DeleteRequestRecord delRecord = new DeleteRequestRecord(store.getDatabase(), uid, store.getSignatureVerifier());
 				if(delRecord.isBefore(mirroringInfo.mTime))
